some normientte
